container-deployment-webapp/requirements.txt
Flask
azure-identity
azure-mgmt-containerinstance
gunicorn
requests